YCbCr JPEG / mask improvements
RGB and masks are kept separate to allow different resampling methods
(lanczos vs. nearest neighbor) to be used and to minimize artifacts.

This isn't straightforward, but is necessary in order to use highly
compressed JPEG-in-TIFFs.

Overviews are now internal (possible due to the decreased file size),
but masks are kept external so that overviews can be created separately
for RGB / mask bands using different resampling methods.

A 4-band VRT is provided when using data in its source projection in
tools that don't fully understand masks (e.g. QGIS,
https://hub.qgis.org/issues/3348) as RGBA (vs. RGB + mask).

Helpful resources:

http://blog.cleverelephant.ca/2015/02/geotiff-compression-for-dummies.html
https://trac.osgeo.org/gdal/wiki/rfc15_nodatabitmask
http://erouault.blogspot.fr/2014/04/advanced-jpeg-in-tiff-uses-in-gdal.html
https://lists.osgeo.org/pipermail/gdal-dev/2014-August/039907.html
https://lists.osgeo.org/pipermail/gdal-dev/2013-July/036682.html
(thread)