Update Lambda dependencies
This uses atimes to determine which files are used in the process of
serving tiles (+ trial and error to get the Lambda-specific ones), which
reduces the deployment size dramatically (and much more simply/reliably
than the previous method of guessing and blacklisting).

track.Dockerfile creates an image that can be used for serving tiles and
determining what's used.

This also adds libjpeg-turbo-devel so that GDAL can deal with JPEG
compression.