Support for internal masks
This uses the materialized alpha band that rasterio provides when a mask
is available (internal or external) and no nodata value has been
provided. As a result, external masks no longer need to be read
explicitly and the slowdown present when using `masked=True` is
bypassed.

This also removes the use of boundless reads (dropped for WarpedVRTs in
rasterio@1.0b2), upgrades rasterio, and drops the scipy dependency that
was being used in an attempt to clean up masked output.